fix: Config 파일 이름 변경 및 기타 클래스 변경 사항
- JwtAuthenticationFIlter : shouldNotFilter URI 변경
- lombok.config : 주석 추가
