+++ lombok.config
@@ -1,2 +1,5 @@
+# JaCoCo에서 lombok이 자동 생성한 코드를 커버리지로 집계하지 않는다.
 lombok.addLombokGeneratedAnnotation = true
+
+# Value 어노테이션이 달린 private final 필드에 값을 주입할 수 있다.
 lombok.copyableAnnotations += org.springframework.beans.factory.annotation.Value
